Refactored KM and GKBF, so that the common architecture is shared.
