+++ src/dymad/models/__init__.py
@@ -1,5 +1,6 @@
 # Import all models
 from dymad.models.model_base import ModelBase
+from dymad.models.model_temp_ucat import ModelTempUCat, ModelTempUCatGraph
 from dymad.models.kbf import DGKBF, DKBF, KBF, GKBF
 from dymad.models.km import DKM, DKMSK, KM
 from dymad.models.ldm import DGLDM, DLDM, GLDM, LDM
@@ -13,6 +14,8 @@ __all__ = [
     "DKMSK",
     "DLDM",
     "ModelBase",
+    "ModelTempUCat",
+    "ModelTempUCatGraph",
     "GKBF",
     "GLDM",
     "KBF",
